Move logic to set instance fields to dedicated instance manager
Instead of setting fields in the task worker handler, move the logic to
set the github org & repo to a dedicated instance manager. This is
necessary to be able to `create()` the objects with the auto validation
happening upon save, which otherwise fails.

Also handle logging on instance objects which haven't been saved yet,
to allow to use some of the instance methods in the manager create
method.