Fix infinite `update_status` loop when going from 'started' to 'booted'
This would happen when the VM boots fast enough for the status to update
from 'started' to 'booted' within a single run of `update_status()`,
skipping the 'active' status in `run_provisioning()`.

This also add support for checking against multiple status in
`sleep_until_status()`, which now accepts lists of status as the
argument.