Handle and display errors from ansible runs
Implemented support for handling error codes and stderr output from ansible
runs. Also included is the support for retries (default is 3) when a run
fails and a new field `progress` on Server model to keep track of progress
of the current status.

The UI was also updated to display prominently the errors happened during
the deployment.