[WIP] Implement strategy pattern
Enable the application user to vary the strategy used for deployment
an instance.

This is done through the instance application config, meaning you will
have to create your own AppConfig, overwrite the strategy attribute
and make sure the new value (a model) implements all methods missing
on the abstract class Server (which the strategy model must inherit
from).

Included is a DigitalOcean strategy. This is a work in progress, we
still need tests for the new strategy in order to reach min coverage
requirement.