Support external mysql and mongo databases
Adds support for instances to use external mysql and mongo databases.
The goal is to preserve databases during sandbox reprovisioning.

Default databases for new instances can be set through the
INSTANCE_MYSQL_URL and INSTANCE_MONGO_URL settings.

If the default is to use external databases but the PR submitter would
like to use local databases, she can include (ephemeral databases) in
the pull request body on the same line as the sandbox url to override
the default.