Avoid mysql security warning in tests.
The mysql command prints a security warning when you pass the password
using the `--password` parameter.
It does not print a security warning if you pass the password using
the MYSQL_PWD environment variable.

Both methods are insecure, but we don't care about that in this case,
since this is only used as a convenience method in tests.