Add arrows to appserver accordion headings.
To make it more obvious that the sections are collapsable/expandable.

This also changes the view so that accordion starts in fully collapsed
mode. This makes it easier to quickly navigate to the desired section.
I find that I am much more often interested in the logs section than
the 'Status' or 'Configuration' sections.