Add support for custom instance domains.
This adds three new fields to instance configuration:

- external_lms_domain
- external_lms_preview_domain
- external_studio_domain

When set, the external domains are used instead of the internal
(opencraft-controlled) domains in ansible configuration and the UI.

Regardless of the presence of external domains, each instance is still
assigned an internal domain and the DNS entry for the internal domain
is always updated to point to the current active appserver.

Customers who use external domains will have to set a CNAME record that
points to the main internal domain.