Set ansible variables for postfix_queue role.
When external SMTP host is defined in settings, populate
POSTFIX_QUEUE_* ansible vars when provisioning app servers.

In addition to merely passing the SMTP server settings to postfix, we
also set it up to rewrite the sender address to use a fixed domain
defined in the settings, that has been pre-validated and authorized as
a sending domain in AuthSMTP.
The local part of the rewritten address (part before @) is the instance
domain, so the rewritten From address looks something like:

pr123.opencraft.hosting@opencraft.hosting

or when using external LMS domain:

courses.mydomain.com@opencraft.hosting.

The original sender address is copied into the Reply-To header.