Only install Django Debug Toolbar when in debugging mode.
The toolbar mostly disables itself when Django is not in debugging mode, but it
still leaks some memory in that case.  In particular, the LoggingCollector
collects all log messages in memory unconditionally, leading to a small but
continuous memory leak.  The debug_toolbar app is also installed on all Huey
workers, where it is essentially useless even when in debug mode.