[BB-835] Use a sub-domain instead of a prefixed domain for Studio, Preview etc. (#385)
* Use a sub-domain instead of a prefix domain for Studio, Preview etc.

From the Ironwood release, it is a requirement to serve the LMS and
Studio from cookie-compatible domains. So the studio domain is now
a sub-domain of the LMS domain. To allow session sharing, the
'EDXAPP_SESSION_COOKIE_DOMAIN' ansible variable has been set to
'.<LMS domain>'. For consistency, all the other domains - preview,
discovery, ecommerce are updated to be sub-domains of the LMS domain.

This also adds a new ansible variable
'EDXAPP_LOGIN_REDIRECT_WHITELIST' which allows the redirect to Studio
after logging in via the LMS.