all values stored in consul are json serialized
An exception occurs if the caller provides a non json serializable
value.

In the context of Ocim (or the consumers of the information written by
Ocim in consul) there is no use case where data that is non json
serializable (but converted into string) would be useful.

It makes the get/put operations consistent instead of inconsistent in
some cases. For instance:

  - put('k', None) => get('k') returns None instead of the string
    "None"
  - put('k', 'null') => get('k') returns "null" instead of None