allow to differentiate a non-existing key from key with value None
When asked for the index, the value '0' is returned if the key does
not exist and the value is set to None. It can then be distinguished
from an existing object which contains a json serialized object equal
to None because such an object will be returned with an index greater
than zero.