pylint: disable unidiomatic-typecheck
The identity check (is)[0] has a semantic that is different from
isinstance[1]. There may be cases where it is more idiomatic to use
isinstance rather than is to compare types. But these cases are not
explained[2] and the pylint check incorrectly assumes they can be used
interchangeably in some cases. In order to avoid misleading false
positives that are confusing to the developper, it is better to
disable this warning entirely.

[0] https://docs.python.org/3.6/reference/expressions.html#is
[1] https://docs.python.org/3.6/library/functions.html#isinstance
[2] $ pylint --help-msg=unidiomatic-typecheck
:unidiomatic-typecheck (C0123): *Using type() instead of isinstance() for a typecheck.*
  The idiomatic way to perform an explicit typecheck in Python is to use
  isinstance(x, Y) rather than type(x) == Y, type(x) is Y. Though there are
  unusual situations where these give different results. This message belongs to
  the basic checker.