Disable broker heartbeat for celery workers
Celery workers are dodgy and don't appear to be actually sending
heartbeats even when configured (as by default) to a non-zero heartbeat
interval. This instructs them to disable amqp protocol heartbeats
altogether. This, combined with setting heartbeat to 0 on the rabbitmq
server, disables amqp protocol heartbeat checks, and we should avoid
getting any connection reset issues.