Make OCIM tests independent of environment locale
This patch affects the following tests:
- test_string_to_file_path_ok
 - test_get_pr_by_number
 - test_get_pr_info_by_number

These tests currently break if 2 conditions occur simultaneously:

- they are run in Python 3.6-, which always uses the locally configured locale by default
- the environment locale is the (default) legacy C locale, which uses ASCII encoding

On the other hand, in Python 3.7+, this problem does not happen. If the environment is
using a legacy C locale, Python coerces it to a UTF-8 based locale and avoids the issue
by implementing https://www.python.org/dev/peps/pep-0538/ , as documented in
https://docs.python.org/3/whatsnew/3.7.html.

Unfortunately, the OCIM tests currently run in a Vagrant VM with Ubuntu Xenial and
Python 3.6.9, so they are subject to the problem above. In order to avoid this possibility,
these tests now explicitly set encoding to UTF-8 before reading/writing to a file.