Use sum instead of single value for NR alerts
This should allow the alerts to be auto closed.

With the original single_value setup,
alerts would only be autoclosed if _every_ query returned at least 1
for a full 16 minutes.
This is impossible, because ping monitors ping in only 5 minute
intervals, and the range for a single query is 1 minute,
so there will be at most 2 successive queries returning at least 1.

By switching to sum, it has the same behaviour of opening an alert,
but the auto close condition becomes 'at least 1 in the last 16 minutes',
which is an easy condition to meet once the instance comes back online.