Updates NewRelic policy to check for failure instead
This check checks for failure instead of success.

This check also makes sure that 3 consecutive failures happened before creating an alert/violation.

Once 3 alerts happen in the matter of 16 minutes, a violation will be opened. In order to guarantee the alert is closed, we use the 'loss of signal' or 'signal expiration'. If we don't have any signal for a duration of 16 mins, ie no failed request for 16 mins, it means that all issues were resolved, so close the violation.