[SE-3953] Updates NewRelic policy to check for failure instead (#690)
This check checks for failure instead of success.

This check also makes sure that 2 consecutive failures happened before creating an alert/violation.

Once 2 alerts happen in the matter of 11 minutes, a violation will be opened. In order to guarantee the alert is closed, we use the 'loss of signal' or 'signal expiration'. If we don't have any signal for a duration of 11 mins, ie no failed request for 11 mins, it means that all issues were resolved, so close the violation.

This change is because New Relic isn't properly sending requests. Check the New Relic Support ticket on SE-3953 for more information.