Set an expiry time on the DNS operation lock.
We use a global lock to ensure that only one process at a time is
modifying DNS records.

By default the lock never expires, which means that if Ocim crashes
before being able to release the lock, DNS operations contiinue being
locked indefinitely even when Ocim comes back up.

We need to put a timeout on the DNS operation to avoid this from
happening.